pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xray-monitor"
version = "10.0.0"
description = "TUI monitor for personal Xray VPN server"
requires-python = ">=3.9"
dependencies = [
    "textual>=0.47",
    "grpcio",
    "protobuf",
    "psutil",
    "qrcode",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.0"]

[project.scripts]
xray-monitor = "xray_monitor.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/xray_monitor"]
